GH-195: General clean-up
